use specific bookcard for search book page
